Data type bug fix

+++ src/main/java/core/basesyntax/strategy/impl/BirthdayDiscountService.java
@@ -3,7 +3,7 @@ package core.basesyntax.strategy.impl;
 import core.basesyntax.strategy.DiscountService;
 
 public class BirthdayDiscountService implements DiscountService {
-    private static final int DISCOUNT = 33;
+    private static final double DISCOUNT = 33;
 
     @Override
     public double getDiscount() {
