+++ src/main/java/core/basesyntax/strategy/impl/BlackFridayDiscountService.java
@@ -3,7 +3,7 @@ package core.basesyntax.strategy.impl;
 import core.basesyntax.strategy.DiscountService;
 
 public class BlackFridayDiscountService implements DiscountService {
-    private static final int DISCOUNT = 45;
+    private static final double DISCOUNT = 45;
 
     @Override
     public double getDiscount() {
